Mastermind
In this file, I have implemented a version of the Mastermind game in Haskell. In this game, the player must guess a secret sequence of 4 random numbers with unlimited attempts. Additionally is indicated how many correct values are in the correct position and how many correct values are in the wrong position.