Bingo
In this file, I have implemented a version of the Bingo game in Haskell. In this game, numbers between 1 and 90 are drawn randomly, with no repeated numbers shown. The game ends after 90 numbers have been generated, that is, when there are no more possible numbers to be generated without being repeated.